feat: add withPlaceholderScreen helper
